in utils add functions create element and render element, delete create-elem, render-elem
