--- js/create-elem.js
@@ -1,7 +1,0 @@
-const createElement = (str) => {
-  const parser = document.createElement(`template`);
-  parser.innerHTML = str;
-  return parser.content;
-};
-
-export default createElement;
