--- js/render-screen.js
@@ -1,8 +1,0 @@
-const mainScreen = document.querySelector(`.main`);
-
-const renderScreen = (screenTemplate) => {
-  mainScreen.innerHTML = ``;
-  mainScreen.appendChild(screenTemplate.cloneNode(true));
-};
-
-export default renderScreen;
