+++ js/templates/renderHeader.js
@@ -1,9 +1,5 @@
 const countNotes = (state) => {
-  return new Array(state.notes).fill(`<img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">`).join(``);
+  return new Array(3 - state.notes).fill(`<img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">`).join(``);
 };
 
-const renderNotes = (state) => {
-  return `${countNotes(state)}`;
-};
-
-export {renderNotes};
+export default countNotes;
